pyproject.toml
[project]
name = "django-todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-filter>=24.3",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
]

[tool.ruff.lint]
select = ["I", "C", "W", "E", "F", "PL"]
ignore = ["E501"]

[tool.ruff]
extend-exclude = ["migrations"]
